102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,13 +10,13 @@ int main(void)
 {
 	int a = 50;
 	int b;
-	unsigned long long first = 1, second = 2, next;
+	unsigned long first = 1, second = 2, next;
 
-	printf("%llu, %llu", first, second);
+	printf("%lu, %lu", first, second);
 	for (b = 2; b < a; b++)
 	{
 		next = first + second;
-		printf(", %llu", next);
+		printf(", %lu", next);
 		first = second;
 		second = next;
 	}
